pyproject.toml
[project]
name = "pywa-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.13.1",
    "pywa>=3.0.0",
    "fastmcp>=0.1.0",
    "python-dotenv",
]
